requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==11.0.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
openapi-codec==1.3.2
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
